changed snake ID to int type

+++ Snake/GameModel/GameModel.cs
@@ -242,7 +242,7 @@ namespace GameModel
         /// note that this is different from UpdateWalls and UpdatePowerups
         /// in that this takes in an already parsed Snake object.
         /// </summary>
-        public bool AddSnake(string playerName, long ID)
+        public bool AddSnake(string playerName, int ID)
         {
             // add a new snake on connection that has name field and ID field provided.
             if (!snakes.ContainsKey(ID))
@@ -377,9 +377,9 @@ namespace GameModel
             growth = 12;
         }
 
-        public Snake(string Name, long iD)
+        public Snake(string Name, int iD)
         {
-            ID = (int)iD;
+            ID = iD;
             name = Name;
             body = new();   // remember to randomize
             dir = new();    // remember to randomize
